requirements.txt
regex
numpy
PyStemmer
pytest
